+++ apps/example-app/app/theme/colors.ts
@@ -8,7 +8,7 @@ const palette = {
   neutral500: "#978F8A",
   neutral600: "#564E4A",
   neutral700: "#3C3836",
-  neutral800: "#191015",
+  neutral800: "#1E1B1A",
   neutral900: "#000000",
 
   primary100: "#F4E0D9",
@@ -51,7 +51,7 @@ export const colors = {
   /**
    * The default text color in many components.
    */
-  text: palette.neutral800,
+  text: palette.neutral200,
   /**
    * Secondary text information.
    */
@@ -59,7 +59,7 @@ export const colors = {
   /**
    * The default color of the screen background.
    */
-  background: palette.neutral200,
+  background: palette.neutral800,
   /**
    * The default border color.
    */
@@ -81,4 +81,7 @@ export const colors = {
    *
    */
   errorBackground: palette.angry100,
+
+  white: palette.neutral100,
+  repoText: palette.secondary300,
 }