+++ apps/example-app/jest.config.js
@@ -11,9 +11,12 @@ module.exports = {
   testPathIgnorePatterns: ["<rootDir>/node_modules/", "<rootDir>/.maestro/", "@react-native"],
   testEnvironment: "jsdom",
   setupFiles: ["<rootDir>/test/setup.ts"],
-  transform:{
-    '^.+\\.test.tsx?$': ['ts-jest', {
-      tsconfig: '<rootDir>/test/test-tsconfig.json'
-    }]
-  }
+  transform: {
+    "^.+\\.test.tsx?$": [
+      "ts-jest",
+      {
+        tsconfig: "<rootDir>/test/test-tsconfig.json",
+      },
+    ],
+  },
 }