+++ lib/reactotron-apisauce/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [3.0.4](https://github.com/infinitered/reactotron/compare/reactotron-apisauce@3.0.3...reactotron-apisauce@3.0.4) (2024-02-14)
+
+## [3.0.3](https://github.com/infinitered/reactotron/compare/reactotron-apisauce@3.0.2...reactotron-apisauce@3.0.3) (2024-02-14)
+
 ### [3.0.2](https://github.com/infinitered/reactotron/compare/reactotron-apisauce@3.0.1...reactotron-apisauce@3.0.2) (2024-01-25)