+++ lib/reactotron-core-server/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [3.1.3](https://github.com/infinitered/reactotron/compare/reactotron-core-server@3.1.2...reactotron-core-server@3.1.3) (2024-02-14)
+
+## [3.1.2](https://github.com/infinitered/reactotron/compare/reactotron-core-server@3.1.1...reactotron-core-server@3.1.2) (2024-02-14)
+
 ### [3.1.1](https://github.com/infinitered/reactotron/compare/reactotron-core-server@3.1.0...reactotron-core-server@3.1.1) (2024-01-25)