+++ lib/reactotron-react-native/CHANGELOG.md
@@ -2,12 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-### [5.0.5](https://github.com/infinitered/reactotron/compare/reactotron-react-native@5.0.4...reactotron-react-native@5.0.5) (2024-01-25)
+### [5.1.3](https://github.com/infinitered/reactotron/compare/reactotron-react-native@5.1.2...reactotron-react-native@5.1.3) (2024-02-15)
 
+### [5.1.2](https://github.com/infinitered/reactotron/compare/reactotron-react-native@5.1.1...reactotron-react-native@5.1.2) (2024-02-14)
+
+## [5.1.1](https://github.com/infinitered/reactotron/compare/reactotron-react-native@5.1.0...reactotron-react-native@5.1.1) (2024-02-14)
+
+## [5.1.0](https://github.com/infinitered/reactotron/compare/reactotron-react-native@5.0.5...reactotron-react-native@5.1.0) (2024-02-07)
+
+
+### Features
+
+* **example-app:** Add a LOT of functionality to the demo app ([#1424](https://github.com/infinitered/reactotron/issues/1424)) [skip ci] ([41a539c](https://github.com/infinitered/reactotron/commit/41a539c7e56e968a83ca61b100468f5267865ddd))
+
+### [5.0.5](https://github.com/infinitered/reactotron/compare/reactotron-react-native@5.0.4...reactotron-react-native@5.0.5) (2024-01-25)
 
 ### Bug Fixes
 
-* **architecture:** Attempt to mitigate compiler error: `Error: Default condition should be last one` ([#1423](https://github.com/infinitered/reactotron/issues/1423)) ([2c7e678](https://github.com/infinitered/reactotron/commit/2c7e678e5afaea79cd01f4ab6e90bd67339fc80a))
+- **architecture:** Attempt to mitigate compiler error: `Error: Default condition should be last one` ([#1423](https://github.com/infinitered/reactotron/issues/1423)) ([2c7e678](https://github.com/infinitered/reactotron/commit/2c7e678e5afaea79cd01f4ab6e90bd67339fc80a))
 
 ### [5.0.4](https://github.com/infinitered/reactotron/compare/reactotron-react-native@5.0.4-beta.19...reactotron-react-native@5.0.4) (2023-12-05)