+++ lib/reactotron-redux/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [3.1.9](https://github.com/infinitered/reactotron/compare/reactotron-redux@3.1.8...reactotron-redux@3.1.9) (2024-02-14)
+
+## [3.1.8](https://github.com/infinitered/reactotron/compare/reactotron-redux@3.1.7...reactotron-redux@3.1.8) (2024-02-14)
+
 ### [3.1.7](https://github.com/infinitered/reactotron/compare/reactotron-redux@3.1.6...reactotron-redux@3.1.7) (2024-01-25)